Renamed '/hooks/transition' directory into '/hooks/navigation'
